Restructure project for simplicity & ease
Restructured the project to provide a simple development template. The
different GUI designers added unneeded complexity because they couldn't
mix/match.

This new structure is basic and should work in any IDE and build a jar
with all dependencies that can be turned in. And importantly, MySQL
connector to interact with the database.

All development now happens in src/main/java.